google-adk-demo/pyproject.toml
[project]
name = "google-adk-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "langsmith>=0.4.38",
    "openinference-instrumentation-google-adk>=0.1.0",
]
